Optimize entity model query performance (#96)
Lazy loading of associations between entities is a well established best practice in JPA. Its main goal is to retrieve only the requested entities from the database and load the related entities only if needed.

The use of FetchType.EAGER  for associations mapping is one of the most common reasons for performance problems, because Hibernate loads eagerly fetched associations when it loads an entity. E.g., when Hibernate loads a Task entity, it also fetches the associated ProcessInstance, Variables, TaskCandidateUsers and TaskCandidateGroups entities. That requires an additional query for each associations and creates dozens or even hundreds of additional queries. This approach is very inefficient and it gets even worse when you consider that Hibernate does that whether or not you will use the association. 

This PR makes explicit use of FetchType.LAZY for all to-many associations. It delays the initialization of the relationship until we use it in business code and avoids a lot of unnecessary queries and improves the performance of query application. It also prevents default eager fetching  of all to-one associations, when loading multiple entities and each of them specify a few of these associations.

It is also enables sensible database indexes  to improve search performance.

Another improvement is to use lazy variable value de-serialization on first property access