Serialize BigDecimals as String
This will ensure that BigDecimals are serialized as String and not as a number, meaning
that double quotes will be added around the value. Serializing it as a number it's problematic
because, by default, it will be deserialized back to Java as double and it will loose precision.
For instance, `1.00` (scale 2) will become `1.0` (scale 1) that are considered as different
values in BigDecimal. By adding the quotes, it will be deserialized as String, but it will not
loose the information about the scale, so it can be easily converted back to BigDecimal.