GH-3694 Pin rabbitmq testcontainer docker image to version `3.8.9` (#456)
* AAE-5592 FIX: fixed the rabbitmq docker version to rabbitmq:3.7.25-management-alpine

The build started failing today due to an update pushed on dockerhub by
rabbitmq. To fix this issue I pinned the version of rabbitmq to
rabbitmq:3.7.25-management-alpine.
I took the occasion to upgrade testcontainers to ver 1.16.0

* AAE-5592 FIX: rabbitmq docker version=rabbitmq:3.9.8-management-alpine

Aligining the version with the one used in production.

* GH-3694 FIX: using rabbitmq docker ver=rabbitmq:3.9.8-management-alpine

Modified other two references to the image name in the code.
Used the RabbitMQContainerApplicationInitializer.class when possible.