GH-3688 Implemented AlfrescoPageRequest.withPage behavior
The method withPage has been introduced by the new spring boot version. The previous implementation was not taking into account the definition of skipCount and maxItem of the AlfrescoPageRequest class itself and was erroneously relying on the delegated Pageable class.

The method has been implemented calculating the page offset by multiplying the page size times the desired page number.