Upgrade to Spring Boot 2.5.2 and Spring Cloud 2020.0.3 (#455)
* GH-3688 Bumbed a new version of spring-cloud and spring-boot

To test the activiti-cloud build in isolation I used the preview
mechanism specifying to use the Activiti preview version:

0.0.1-dev-gicappa-GH-3688-preview-211-SNAPSHOT

* GH-3688 Bumping a new spring-boot/cloud version

Found some minor incompatibilities due to the evolution of the spring-boot/data platform from the 2.4.x to the 2.5.x

- The Pageable interface has a new method withPage() that has been implemented by delegating to the inner peageable implementation.

- QueryRepositoryConfig and RepositoryRestConfig are inheriting from RepositoryRestConfigurer that changed the signature of the method configureRepositoryRestConfiguration() adding the CorsRegistry parameter. The signature has been updated but the parameter has not been taken into consideration.

- In ModelingRestExceptionHandler the DefaultErrorAttributes class changed the signature of getErrorAttributes() switching from a boolean param to include the stack trace to a class ErrorAttributeOptions to manage several kind of error attributes to be included.

* GH-3688 Bumping a new spring-boot/cloud version

FIX: with the new spring-boot version the test class QueryApplicationEntityIT was failing reporting a missing Id in the list of ApplicationEntity returned by the test.
The QueryRepositoryConfig class, the the ApplicationEntity was not listed with the classes allowed to expose their ids.

* GH-3688 Bumping a new spring-boot/cloud version

FIX: with the new spring-boot version the test class QueryAdminProcessServiceTasksIT was failing on the shouldGetServiceTaskById.
In the QueryRepositoryConfig class, the ServiceTaskEntity was not listed with the classes allowed to expose their ids.

* FIX: database initialization in sb 2.5.2 has changed causing a bug in the tests.

It has been added the properties spring.jpa.defer-datasource-initialization=true to the application-test.properties file to be retro-compatible with the previous version.

reference https://github.com/spring-projects/spring-boot/wiki/Spring-Boot-2.5-Release-Notes#hibernate-and-datasql

* Bump of the new released of spring-boot 2.5.3

* FIX: spring-integration-amqp raises an exception creating appcontext

After upgrading spring-boot from 2.5.2 to 2.5.3, the
build started failing because it was not anymore able
to create the ApplicationContext.

The error reported was:

Caused by: org.springframework.context.ApplicationContextException: Failed to start bean 'inputBindingLifecycle'; nested exception is java.lang.NoSuchMethodError: org.springframework.integration.amqp.inbound.AmqpInboundChannelAdapter.<init>(Lorg/springframework/amqp/rabbit/listener/AbstractMessageListenerContainer;)V
	at org.springframework.context.support.DefaultLifecycleProcessor.doStart(DefaultLifecycleProcessor.java:181)

The issue seems to be caused by this change in spring-integration-amqp:
https://github.com/spring-projects/spring-integration/commit/76f77ccd2c5f94e000495357e66c4f99fffc6010

Since the issue happens only due to the dependencies imported for testing it has been fixed rolling
back the library spring-integration-amqp the the previous version (5.5.1)

* Factorized the fix introduced in the spring-boot version 2.5.3

* GH-3688 FIX wrong position of a field / static code analysis

* GH-3688 Aligning the test class to JUnit5

Minor refactorings to align to JUnit5 standards and to improve consistency through the code.

* GH-3688 Implemented AlfrescoPageRequest.withPage behavior

The method withPage has been introduced by the new spring boot version. The previous implementation was not taking into account the definition of skipCount and maxItem of the AlfrescoPageRequest class itself and was erroneously relying on the delegated Pageable class.

The method has been implemented calculating the page offset by multiplying the page size times the desired page number.

* GH-3688 Factorised the rollback of the spring-integration version to 5.5.1

The dependency to version 5.5.1 was introduced twice since the
activiti-cloud-runtime-bundle-service/pom.xml was receiving the rollback
of the spring-integration-amqp to the 5.5.1 version from the
activiti-cloud-build-dependencies-parent/pom.xml.
Overriding the whole spring-integration version using the bom in the
activiti-cloud-build-dependencies-parent/pom.xml solved the issue
(thanks to Elias for the hint).