AAE-23532 Fix race condition for service tasks using connectors (#1487)
Having `MQServiceTaskBehavior ` as an `ActivityBehavior` was causing racing conditions because it's instantiated as a singleton. Given the fact that `MQServiceTaskBehavior` is always wrapped inside a `ServiceTaskDelegateExpressionActivityBehavior` that is already an `ActivityBehavior`, there is no reason to have `MQServiceTaskBehavior` also implementing `ActivityBehavior`. This PR is makes `MQServiceTaskBehavior` implement the new interface `DelegateExecutionFunction` instead of `ActivityBehavior`.